V2/requirements.txt
streamlit
jinja2
python-dotenv
requests